move MediaService to LiveChatService from ChatDataProvider, fix that sent image is not appearing
